+++ packages/core/src/definitions/index.ts
@@ -1,3 +1,4 @@
+export { CheckStatus, AmlReason } from './aml';
 export { AssetUrl, AssetType, AssetCategory } from './asset';
 export type { Asset } from './asset';
 export { AuthUrl, AuthWalletType } from './auth';
@@ -29,6 +30,7 @@ export {
   KycState,
   LimitPeriod,
   KycStepName,
+  KycStepCancelable,
   KycStepType,
   KycStepStatus,
   UrlType,
@@ -139,7 +141,9 @@ export {
   SupportIssueType,
   SupportIssueReason,
   SupportIssueState,
+  SupportIssueInternalState,
   SupportMessageStatus,
+  Department,
 } from './support';
 export type {
   Reaction,