+++ packages/core/src/definitions/transaction.ts
@@ -72,6 +72,7 @@ export enum TransactionState {
   RETURNED = 'Returned',
   UNASSIGNED = 'Unassigned',
   WAITING_FOR_PAYMENT = 'WaitingForPayment',
+  STOPPED = 'Stopped',
 }
 
 export enum TransactionFailureReason {