+++ packages/react/src/definitions/kyc.ts
@@ -8,6 +8,7 @@ export {
   KycState,
   LimitPeriod,
   KycStepName,
+  KycStepCancelable,
   KycStepType,
   KycStepStatus,
   UrlType,