+++ packages/react/src/index.ts
@@ -34,6 +34,7 @@ export { useSwap } from './hooks/swap.hook';
 export { useSupportChat } from './hooks/support.hook';
 
 // Definitions
+export { CheckStatus, AmlReason } from './definitions/aml';
 export { Asset, AssetType, AssetCategory } from './definitions/asset';
 export { BankAccount } from './definitions/bank-account';
 export { Bank } from './definitions/bank';
@@ -88,6 +89,7 @@ export {
   KycInfo,
   KycSession,
   KycStepName,
+  KycStepCancelable,
   KycStepType,
   KycStepStatus,
   UrlType,
@@ -204,7 +206,9 @@ export {
   BlobContent,
   DataFile,
   SupportIssueState,
+  SupportIssueInternalState,
   SupportMessageStatus,
+  Department,
   Reaction,
 } from './definitions/support';